requirements.txt
streamlit==1.32.0
streamlit-drawable-canvas==0.9.3
Pillow
